Corrected some mistakes in the initial spectrum of L waves.
Now it is calculated locally instead of being read from an external file.
I also have changed the length of the VQQ vector from 10 to 6.

The next text will involve the time evolution with the initial spectrum including the electrostatic bremsstrahlung and collisional damping effects in the Langmuir and ion-sound spectra. This feature makes this test different from the test in "tevol-double" that does not include the new effects in the initial spectrum.

Let's see what happens then! =)